:recycle: refactor: ErrorCode 인터페이스 사용
