+++ src/main/java/sandbox/semo/security/authentication/MemberAuthFailureHandler.java
@@ -10,7 +10,7 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 import org.springframework.stereotype.Component;
-import sandbox.semo.security.exception.ErrorCode;
+import sandbox.semo.security.exception.AuthErrorCode;
 import sandbox.semo.security.util.JsonResponseHelper;
 
 @Log4j2
@@ -21,16 +21,16 @@ public class MemberAuthFailureHandler implements AuthenticationFailureHandler {
     public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
             AuthenticationException exception) throws IOException, ServletException {
         log.error(">>> [ ❌ 인증 실패: {} ]", exception.getMessage());
-        ErrorCode errorCode = getErrorMessage(exception);
-        JsonResponseHelper.sendJsonErrorResponse(response, errorCode);
+        AuthErrorCode authErrorCode = getErrorMessage(exception);
+        JsonResponseHelper.sendJsonErrorResponse(response, authErrorCode);
     }
 
-    private ErrorCode getErrorMessage(AuthenticationException exception) {
+    private AuthErrorCode getErrorMessage(AuthenticationException exception) {
         if (exception instanceof UsernameNotFoundException
                 || exception instanceof BadCredentialsException) {
-            return ErrorCode.INVALID_CREDENTIALS;
+            return AuthErrorCode.INVALID_CREDENTIALS;
         }
-        return ErrorCode.UNAUTHORIZED_USER;
+        return AuthErrorCode.UNAUTHORIZED_USER;
     }
 
 }
